backend/requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
langchain-classic
requests
httpx
python-dotenv
rich
pydantic
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
